Upgrade To Rails 7
Upgrades the app to rails version 7 which was performed by blowing away
all code and copying back previous files for the oauth flow.

There were some minor changes such as switching from secrets to
credentials in which the master key is committed for this repo as this
is for demo purposes only.